requirements.txt
asgiref==3.7.2
branca==0.7.1
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
crispy-tailwind==0.5.0
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0
django-appconf==1.0.6
django-compressor==4.4
django-crispy-forms==2.1
django-import-export==3.3.5
django-ranged-response==0.2.0
django-recaptcha==4.0.0
django-simple-captcha==0.6.0
django-webpack-loader==3.0.1
et-xmlfile==1.1.0
folium==0.15.1
future==0.18.3
geocoder==1.38.1
idna==3.6
Jinja2==3.1.3
MarkupPy==1.14
MarkupSafe==2.1.5
numpy==1.26.3
odfpy==1.4.1
openpyxl==3.1.2
Pillow==10.1.0
psycopg2==2.9.9
python-dotenv==1.0.0
PyYAML==6.0.1
ratelim==0.1.6
rcssmin==1.1.1
reportlab==4.0.9
requests==2.31.0
rjsmin==1.2.1
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
tzdata==2023.3
urllib3==2.1.0
uuid==1.30
xlrd==2.0.1
xlwt==1.3.0
xyzservices==2023.10.1
